[TASK] Use TYPO3's caches to reduce number of API calls
TYPO3 executes a number of HTTP calls to the TYPO3 API when retrieving
data about TYPO3 releases. The sum of these requests take several seconds
and slow down the system every time the Nagios extension list is
generated.

As the data from the TYPO3 API does not change often, multiple requests
within a short time period are unnecessary.

This change stores the data in TYPO3's local caches and uses the cached
data for a certain time period. This reduces the number of requests to
the API and significantly improves the performance.